Move DropUncle condition into DropUncle component.
